Simplified instructions for CSV file parameters
A slightly more simplified explanation of how to open the import CSV file dialogue box. For users who have just opened straight to this help section, or may be unfamiliar with setting up parameters in Minsky, it could be helpful to specify which menu to create a parameter from and to make the process more explicit.